refactor(drawnix): move toast types to toast module
